Changed index title bar title and fixed a bug in linking apps
Changed the title bar label for both apps to reflect which they are to
avoid confusion.  Also disabled ajax highjacking when switching between
the two apps as jQuery was not cooperating. Copied over the contents of
the about page from the gold app into the bronze app and changed
content as necessary.